pyproject.toml
[tool.poetry]
name = "pushka-bot"
version = "0.1.0"
description = ""
authors = ["Link13-tech <151656048+Link13-tech@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
